wp-now: refactor readFileHead to independent file
