+++ packages/wp-now/src/wp-playground-wordpress/index.ts
@@ -5,3 +5,4 @@ export * from './is-wp-content-directory';
 export * from './is-wordpress-directory';
 export * from './is-wordpress-develop-directory';
 export * from './get-plugin-file';
+export * from './read-file-head';
